changed subsetting order
It is more logical to first create a small table with genes and human homologs, and then to filter on is.na(human homologs), to focus on the is.na function. Otherwise we would have to explicitly explain that we are creating a temporary table to visualise the result of the filtering, and this is seen just after, when introducing the pipes.